+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### v0.4.0
 
 * cp of a single file hangs on a prompt #5 (thanks @roblabla)
+* Add support for custom windows conversion hint
 
 ### v0.3.1 (2019-06-16)